+++ app/Http/Controllers/EncomendasController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Auth;
 use App\Models\Encomenda;
 use App\Models\Produto;
 
@@ -19,8 +21,8 @@ class EncomendasController extends Controller
     }
 
     public function create(){
-        $produto = Produto::all();
-        return view('encomenda.create',['produto'=>$produto]);
+            $produto = Produto::all();
+            return view('encomenda.create',['produto'=>$produto]);
     }
 
     public function store(Request $r){
@@ -38,35 +40,54 @@ class EncomendasController extends Controller
     }
 
     public function delete(Request $r){
-        $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
-        return view('encomenda.delete',['encomenda'=>$encomenda]);
+        if(Gate::allows('admin')){
+            $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
+            return view('encomenda.delete',['encomenda'=>$encomenda]);
+        }
+        else{
+            return redirect()->route('encomenda.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function destroy(Request $r){
-        $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
-        $encomenda->delete();
-        return redirect()->route('encomenda.index');
+        if(Gate::allows('admin')){
+            $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
+            $encomenda->delete();
+            return redirect()->route('encomenda.index');
+        }
+        else{
+            return redirect()->route('encomenda.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function edit(Request $r){
-        $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
-        $produtos = Produto::all();
- //dd($produto);
-        $produtosEncomendas=[];
-        foreach($encomenda->produtos as $produto){
-            $produtosEncomendas[]=$produto->id_produto;
+        if(Gate::allows('admin')){
+            $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
+            $produtos = Produto::all();
+            $produtosEncomendas=[];
+            foreach($encomenda->produtos as $produto){
+                $produtosEncomendas[]=$produto->id_produto;
+            }
+            return view('encomenda.edit',['encomenda'=>$encomenda,'produtoEnc'=>$produtosEncomendas,'produto'=>$produtos]);
         }
-        return view('encomenda.edit',['encomenda'=>$encomenda,'produtoEnc'=>$produtosEncomendas,'produto'=>$produtos]);
+        else{
+            return redirect()->route('encomenda.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function update(Request $r){
-        $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
-        $editarEncomenda = $r->validate([
-            'id_produto' => ['nullable','numeric'],
-            'quantidade' => ['nullable','numeric'],
-            'preco' =>['nullable','numeric'],
-        ]);
-        $encomendaUpdated = $encomenda->update($editarEncomenda);
-        return redirect()->route('encomenda.show',['id'=>$encomenda->id_encomenda]);
+        if(Gate::allows('admin')){
+            $encomenda = Encomenda::where('id_encomenda',$r->id)->first();
+            $editarEncomenda = $r->validate([
+                'id_produto' => ['nullable','numeric'],
+                'quantidade' => ['nullable','numeric'],
+                'preco' =>['nullable','numeric'],
+            ]);
+            $encomendaUpdated = $encomenda->update($editarEncomenda);
+            return redirect()->route('encomenda.show',['id'=>$encomenda->id_encomenda]);
+        }
+        else{
+            return redirect()->route('encomenda.index')->with('msg','Não têm permissão');
+        }
     }
 }
