+++ app/Http/Controllers/VendedoresController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Vendedor;
 use Auth;
 
@@ -20,46 +21,76 @@ class VendedoresController extends Controller
     }
 
     public function create(){
-        return view('vendedor.create');
+        if(Gate::allows('admin')){
+            return view('vendedor.create');
+        }
+        else{
+            return redirect()->route('vendedor.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function store(Request $r){
-        $novoVendedor = $r->validate([
-            'id_vendedor' => ['nullable','numeric','min:1'],
-            'nome' => ['required','min:1','max:50'],
-            'especialidade' => ['required','min:1','max:50'],
-            'email'=> ['nullable','min:1','max:50'],
-        ]);
+        if(Gate::allows('admin')){
+            $novoVendedor = $r->validate([
+                'id_vendedor' => ['nullable','numeric','min:1'],
+                'nome' => ['required','min:1','max:50'],
+                'especialidade' => ['required','min:1','max:50'],
+                'email'=> ['nullable','min:1','max:50'],
+            ]);
 
-        $vendedor = Vendedor::create($novoVendedor);
-        return redirect()->route('vendedor.show',['id'=>$vendedor->id_vendedor]);
+            $vendedor = Vendedor::create($novoVendedor);
+            return redirect()->route('vendedor.show',['id'=>$vendedor->id_vendedor]);
+        }
+        else{
+            return redirect()->route('vendedor.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function delete(Request $r){
-        $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
-        return view('vendedor.delete',['vendedor'=>$vendedor]);
+        if(Gate::allows('admin')){
+            $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
+            return view('vendedor.delete',['vendedor'=>$vendedor]);
+        }
+        else{
+            return redirect()->route('vendedor.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function destroy(Request $r){
-        $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
-        $vendedor->delete();
-        return redirect()->route('vendedor.index');
+        if(Gate::allows('admin')){
+            $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
+            $vendedor->delete();
+            return redirect()->route('vendedor.index');
+        }
+        else{
+            return redirect()->route('vendedor.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function edit(Request $r){
-        $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
-        return view('vendedor.edit',['vendedor'=>$vendedor]);
+        if(Gate::allows('admin')){
+            $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
+            return view('vendedor.edit',['vendedor'=>$vendedor]);
+        }
+        else{
+            return redirect()->route('vendedor.index')->with('msg','Não têm permissão');
+        }
     }
 
     public function update(Request $r){
-        $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
-        $editarVendedor = $r->validate([
-            'id_vendedor' => ['nullable','numeric','min:1'],
-            'nome' => ['required','min:1','max:50'],
-            'especialidade' => ['required','min:1','max:50'],
-            'email'=> ['nullable','min:1','max:50'],
-        ]);
-        $vendedorAtualizado = $vendedor->update($editarVendedor);
-        return redirect()->route('vendedor.show',['id'=>$vendedor->id_vendedor]);
+        if(Gate::allows('admin')){
+            $vendedor = Vendedor::where('id_vendedor',$r->id)->first();
+            $editarVendedor = $r->validate([
+                'id_vendedor' => ['nullable','numeric','min:1'],
+                'nome' => ['required','min:1','max:50'],
+                'especialidade' => ['required','min:1','max:50'],
+                'email'=> ['nullable','min:1','max:50'],
+            ]);
+            $vendedorAtualizado = $vendedor->update($editarVendedor);
+            return redirect()->route('vendedor.show',['id'=>$vendedor->id_vendedor]);
+        }
+        else{
+            return redirect()->route('vendedor.index')->with('msg','Não têm permissão');
+        }
     }
 }
